+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 **?.?.? Changes:**
 
-* Added 3 new effects:
+* Added 4 new effects:
   * Inventory Swap: Swaps the inventories of each player with another player. Multiplayer only.
   * No Equipment Cooldowns: Removes all equipment cooldowns, lasts 60 seconds.
   * Disable Equipment: Disables all equipment activation, lasts 60 seconds.
+  * All Items Are A Random Item: All items on the stage get turned into a random item. Essentially Artifact of Kin for items. Lasts 1 stage.
 
 * Spawn Void Seed:
   * Void Seed can now spawn at any spot on the stage, not just the ones normally available as Void Seed spawn locations